Add user authentication and thoughts management routes, models, and middleware
