+++ package.json
@@ -22,9 +22,13 @@
     "@babel/core": "^7.17.9",
     "@babel/node": "^7.16.8",
     "@babel/preset-env": "^7.16.11",
+    "bcrypt-nodejs": "^0.0.3",
+    "bcryptjs": "^3.0.2",
     "cors": "^2.8.5",
-    "express": "^4.17.3",
+    "dotenv": "^16.5.0",
+    "express": "^4.21.2",
     "mongoose": "^8.15.1",
-    "nodemon": "^3.0.1"
+    "nodemon": "^3.0.1",
+    "validator": "^13.15.15"
   }
 }
